+++ imports/pages/groups/finder/index.js
@@ -5,7 +5,10 @@ import ReactMixin from "react-mixin";
 import gql from "graphql-tag";
 import { withRouter } from "react-router";
 
-import Split, { Left, Right } from "../../../components/@primitives/layout/split";
+import Split, {
+  Left,
+  Right,
+} from "../../../components/@primitives/layout/split";
 import Headerable from "../../../deprecated/mixins/mixins.Header";
 import { nav as navActions } from "../../../data/store";
 
@@ -14,19 +17,23 @@ import Result from "./Result";
 
 const defaultArray = [];
 class TemplateWithoutData extends Component {
-
   static propTypes = {
     dispatch: PropTypes.func.isRequired,
     router: PropTypes.object.isRequired,
     location: PropTypes.object.isRequired,
+    autofill: PropTypes.object.isRequired,
     attributes: PropTypes.object.isRequired,
     content: PropTypes.object.isRequired,
-  }
+  };
 
   state = {
     tags: [],
     query: null,
-  }
+    latitude: null,
+    longitude: null,
+    campus: "",
+    zip: "",
+  };
 
   componentWillMount() {
     this.props.dispatch(navActions.setLevel("TOP"));
@@ -43,62 +50,152 @@ class TemplateWithoutData extends Component {
     }
   }
 
+  deviceReadyFunction = () => {
+    navigator.geolocation.getCurrentPosition(this.setGeolocation);
+  };
+
+  setGeolocation = (position: Object) => {
+    this.setState({
+      latitude: position.coords.latitude,
+      longitude: position.coords.longitude,
+    });
+  };
+
   getResults = () => {
-    const { tags, query } = this.state;
     const { router, location } = this.props;
+    const { latitude, longitude } = this.state;
+    // create an array of the attributes returned by graphql
+    const attributeTags = this.props.attributes.tags.map(tag => tag.value);
+
+    const tags = [];
+    const query = [];
+
+    // find the tags that aren't defined
+    this.state.tags.forEach(e => {
+      if (e && attributeTags.indexOf(e) > -1) {
+        tags.push(e);
+      } else {
+        query.push(e);
+      }
+    });
 
     if (!location.query) location.query = {};
 
-    if (query) location.query.q = query;
+    if (query.length) location.query.q = query.join(",").toLowerCase();
     if (tags.length) location.query.tags = tags.join(",").toLowerCase();
 
     if (location.query.campuses) delete location.query.campuses;
     if (location.query.schedules) delete location.query.schedules;
 
-    // reset state
-    this.setState({ tags: [], query: null });
+    this.setState({ tags: [], query: null, latitude, longitude });
     router.push(location);
-  }
+  };
 
-  inputOnChange = (value) => {
+  campusOnChange = (c: String) => {
     this.setState({
-      tags: this.state.tags,
-      query: value,
+      campus: c || "",
     });
-  }
+  };
 
-  tagOnClick = (tag) => {
+  inputOnChange = (value: String) => {
+    // split each element of the query string into its own array element
+    // at this point the query string also includes tags
+    const queryString = value.split(/[, ]+/);
+
+    const attributeTags = this.props.attributes.tags.map(tag => tag.value);
+
+    // current state of tags to work off of.
+    const newTags = [...this.state.tags];
+
+    // remove the tags that have been removed from the search field
+    const removeTags = newTags.filter(e => {
+      if (e && queryString.indexOf(e) < 0) {
+        newTags.splice(newTags.indexOf(e), 1);
+      }
+    });
+
+    // map over the querystring elements and push the elements that are found in
+    // the defined list but not currently in state
+    const addTags = queryString.filter((e, i, a) => {
+      if (e === "friendly" && a[i - 1] === "kid") {
+        newTags.push("kid friendly");
+      }
+
+      if (e && newTags.indexOf(e) < 0 && attributeTags.indexOf(e) > -1) {
+        newTags.push(e);
+      }
+    });
+
+    this.setState({
+      tags: newTags,
+      query: value,
+      latitude: this.state.latitude,
+      longitude: this.state.longitude,
+    });
+  };
+
+  tagOnClick = (tag: String) => {
     const tagList = [...this.state.tags];
+    let queryString = this.state.query || "";
+
     if (tagList.indexOf(tag) > -1) {
-      // remove the tag from the list string
+      queryString = queryString.replace(new RegExp(`(,? ?)${tag}`), "");
+      if (queryString[0] === ",") {
+        queryString = queryString.substring(1);
+      }
       tagList.splice(tagList.indexOf(tag), 1);
     } else {
+      queryString =
+        queryString && queryString.length ? `${queryString}, ${tag}` : `${tag}`;
       tagList.push(tag);
     }
 
-    this.setState({ tags: tagList });
-  }
+    this.setState({
+      tags: tagList,
+      query: queryString.trim(),
+      latitude: this.state.latitude,
+      longitude: this.state.longitude,
+    });
+  };
 
-  submitTags = (e) => {
+  submitTags = (e: Event) => {
     if (e) e.preventDefault();
     this.getResults();
-  }
+  };
 
-  findByQuery = (e) => {
+  findByQuery = (e: Event) => {
     if (e) e.preventDefault();
     document.getElementById("search").blur();
     this.getResults();
-  }
+  };
 
   /* eslint-disable max-len */
   render() {
-    const { attributes, location, content } = this.props;
-    if (location.query && (
-      location.query.tags
-      || location.query.q
-      || location.query.campuses
-      || location.query.schedules
-    )) return <Result />;
+    const { attributes, location, content, autofill } = this.props;
+
+    if (
+      location.query &&
+      (location.query.tags ||
+        location.query.q ||
+        location.query.campuses ||
+        location.query.schedules)
+    ) {
+      return <Result />;
+    }
+
+    let selectedCampus = this.state.campus;
+    let zipCode = this.state.zip;
+
+    if (
+      !autofill.loading &&
+      autofill.person &&
+      !this.state.campus &&
+      !this.state.zip
+    ) {
+      selectedCampus = autofill.person.campus.name;
+      zipCode = autofill.person.home.zip;
+    }
+
     return (
       <div>
         <Split>
@@ -111,8 +208,26 @@ class TemplateWithoutData extends Component {
         <Left scroll classes={["background--light-secondary"]}>
           <Layout
             canSearchTags={false || this.state.tags.length || this.state.query}
+            campuses={
+              autofill.loading
+                ? [""]
+                : autofill.campuses
+                    .filter(x => {
+                      if (x.name === "Web") {
+                        return false;
+                      }
+
+                      return true;
+                    })
+                    .map(x => x.name)
+            }
+            selectedCampus={selectedCampus}
+            zip={zipCode}
+            campusOnChange={this.campusOnChange}
+            searchQuery={this.state.query}
             tags={(attributes && attributes.tags) || defaultArray}
             tagOnClick={this.tagOnClick}
+            selectedTags={this.state.tags}
             submitTags={this.submitTags}
             findByQuery={this.findByQuery}
             inputOnChange={this.inputOnChange}
@@ -120,11 +235,32 @@ class TemplateWithoutData extends Component {
           />
         </Left>
       </div>
-
     );
   }
   /* eslint-enable max-len */
 }
+
+const AUTOFILL_META_QUERY = gql`
+  query autoFillMeta {
+    person: currentPerson {
+      id
+      firstName
+      nickName
+      home {
+        zip
+      }
+      campus {
+        name
+      }
+    }
+    campuses {
+      id
+      name
+    }
+  }
+`;
+
+const withAutoFillMeta = graphql(AUTOFILL_META_QUERY, { name: "autofill" });
 
 const GROUP_ATTRIBUTES_QUERY = gql`
   query GetGroupAttributes {
@@ -136,14 +272,20 @@ const GROUP_ATTRIBUTES_QUERY = gql`
   }
 `;
 
-const withGroupAttributes = graphql(GROUP_ATTRIBUTES_QUERY, { name: "attributes" });
+const withGroupAttributes = graphql(GROUP_ATTRIBUTES_QUERY, {
+  name: "attributes",
+});
 
 const TAGGED_CONTENT_QUERY = gql`
-  query GetTaggedContent($tagName: String!, $limit: Int, $includeChannels: [String]) {
+  query GetTaggedContent(
+    $tagName: String!
+    $limit: Int
+    $includeChannels: [String]
+  ) {
     entries: taggedContent(
-      tagName: $tagName,
-      limit: $limit,
-      includeChannels: $includeChannels,
+      tagName: $tagName
+      limit: $limit
+      includeChannels: $includeChannels
       cache: false
     ) {
       entryId: id
@@ -168,29 +310,25 @@ const TAGGED_CONTENT_QUERY = gql`
 
 const withTaggedContent = graphql(TAGGED_CONTENT_QUERY, {
   name: "content",
-  options: ({
+  options: {
     variables: {
       tagName: "community",
       includeChannels: ["articles"],
       limit: 2,
     },
-  }),
+  },
 });
 
-const mapStateToProps = (state) => ({ location: state.routing.location });
+const mapStateToProps = state => ({ location: state.routing.location });
 
 export default withRouter(
   connect(mapStateToProps)(
     withGroupAttributes(
-      withTaggedContent(
-        ReactMixin.decorate(Headerable)(
-          TemplateWithoutData
-        )
-      )
-    )
-  )
+      withAutoFillMeta(
+        withTaggedContent(ReactMixin.decorate(Headerable)(TemplateWithoutData)),
+      ),
+    ),
+  ),
 );
 
-export {
-  TemplateWithoutData,
-};
+export { TemplateWithoutData };